pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embedrag"
version = "0.5.0"
description = "Production-grade read/write split RAG retrieval system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "numpy>=1.26",
    "faiss-cpu>=1.8",
    "msgpack>=1.0",
    "zstandard>=0.22",
    "aiohttp>=3.9",
    "boto3>=1.34",
    "prometheus-client>=0.20",
    "httpx>=0.27",
    "croniter>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
embedrag = "embedrag.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/embedrag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
    "performance: marks performance benchmarks",
]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
